Fix (?) bug where keyboard gets stuck on layer
This usually would happen with the Media layer. It looks like the
logic in the code (testing < 0) assumed that these counters could
go negative where, in reality because they are unsigned, they would
actually wrap around to a really big positive number. Not sure if
this is what was causing the issue because it's difficult to
reproduce...